requirements.txt
altgraph==0.17.4
cffi==1.17.1
cryptography==45.0.3
m2secret-py3==1.3
py2app==0.28.8
pycparser==2.22
pyinstaller
six==1.17.0
pip-upgrader
